+++ examples/context/src/main.rs
@@ -1,6 +1,6 @@
 use sycamore::context::{ContextProvider, ContextProviderProps};
 use sycamore::prelude::*;
-use sycamore::rx::use_context;
+use sycamore::reactive::use_context;
 
 #[component(Counter<G>)]
 fn counter() -> Template<G> {
