+++ packages/sycamore/src/motion.rs
@@ -3,7 +3,7 @@ use std::rc::Rc;
 
 use chrono::{prelude::*, Duration};
 
-use crate::rx::Signal;
+use crate::reactive::Signal;
 use crate::utils::{loop_raf, Task};
 
 /// Describes a trait that can be linearly interpolate between two points.
