+++ packages/sycamore/src/utils/render.rs
@@ -5,7 +5,7 @@ use std::rc::Rc;
 use ahash::AHashMap;
 
 use crate::generic_node::GenericNode;
-use crate::rx::create_effect;
+use crate::reactive::create_effect;
 use crate::template::{Template, TemplateType};
 
 /// Insert a [`GenericNode`] under `parent` at the specified `marker`. If `initial` is `Some(_)`,
